wap_player demo: Move control of the fs handle to the wav player.
This splits the package WAV_Reader in two, the WAV header reading and the
actual playing. The player is now fully in charge of opening and closing files
as it's the only place where we can make absolutely sure that any opened file
is closed (this is for maintenance and readability reason).

Also the support for parsing the wav header files and the control of the audio
stream is quite different, so this made sense to move the two apart.